# Remove StrictData to improve performance