## Description
It looks like `StrictData` considerably worsens performance. Those few changes make the `ethereum-tests` suite run way faster and use less memory.

Before:
```
All 18425 tests passed (59.20s)
 339,633,148,432 bytes allocated in the heap
  50,851,677,224 bytes copied during GC
   5,848,354,664 bytes maximum residency (58 sample(s))
     228,278,424 bytes maximum slop
           16567 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     74371 colls,     0 par   11.506s  11.521s     0.0002s    0.0412s
  Gen  1        58 colls,     0 par   14.590s  14.591s     0.2516s    2.9466s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time   62.712s  ( 62.701s elapsed)
  GC      time   26.096s  ( 26.112s elapsed)
  EXIT    time    0.000s  (  0.000s elapsed)
  Total   time   88.808s  ( 88.814s elapsed)

  %GC     time       0.0%  (0.0% elapsed)

  Alloc rate    5,415,776,764 bytes per MUT second

  Productivity  70.6% of total user, 70.6% of total elapsed
```

After:
```
All 18425 tests passed (44.06s)                                                                                                                                                                                                               
 309,213,114,032 bytes allocated in the heap                                                                                                                                                                                                  
  31,994,007,040 bytes copied during GC                                                                                                                                                                                                       
   3,762,108,336 bytes maximum residency (57 sample(s))                                                                                                                                                                                       
     185,952,336 bytes maximum slop                                                                                                                                                                                                           
            7117 MiB total memory in use (0 MB lost due to fragmentation)                                                                                                                                                                     
                                                                                                                                                                                                                                              
                                     Tot time (elapsed)  Avg pause  Max pause                                                                                                                                                                 
  Gen  0     67440 colls,     0 par    7.469s   7.481s     0.0001s    0.0420s                                                                                                                                                                 
  Gen  1        57 colls,     0 par    8.232s   8.233s     0.1444s    0.9039s                                                                                                                                                                 
                                                                                                                                                                                                                                              
  INIT    time    0.000s  (  0.000s elapsed)                                                                                                                                                                                                  
  MUT     time   58.143s  ( 58.134s elapsed)                                                                                                                                                                                                  
  GC      time   15.701s  ( 15.714s elapsed)                                                                                                                                                                                                  
  EXIT    time    0.000s  (  0.000s elapsed)                                                                                                                                                                                                  
  Total   time   73.844s  ( 73.849s elapsed)                                                                                                                                                                                                  
                                                                                                                                                                                                                                              
  %GC     time       0.0%  (0.0% elapsed)                                                                                                                                                                                                     
                                                                                                                                                                                                                                              
  Alloc rate    5,318,144,343 bytes per MUT second                                                                                                                                                                                            
                                                                                                                                                                                                                                              
  Productivity  78.7% of total user, 78.7% of total elapsed           
```

## Checklist

- [x] tested locally
- [ ] added automated tests
- [ ] updated the docs
- [ ] updated the changelog